Implement Write() and test using fakes
Every write operation consists of four 8bit commands. The first three bytes
are for selecting the register and sub-register to write to, and the last
byte is the actual data for that sub-register.

I've written a fakes package based on the i2c library so that the data that
would be sent to the controller can be inspected by tests.