Implement InterruptMask + InterruptFrame registers
InterruptFrame only has a single option so we don't need an accompanying
struct for the set method.

`InterruptMask.PowerOrReset` defaults on the device to being true but I
haven't represented this in the struct because it would be more complicated.
I assume that someone setting this knows what they're doing and will
explicitly set it if they need it.