Add associate PWM set to Framer interface
So that we can set this on generic frame sizes in `as1130ctl`. Now that the
exported field is gone we can call the get method `PWMSet()`. There's some
slightly annoying repetition in the set method name, but renaming "PWM set"
to "PWM index" wouldn't match the datasheet and might suggest that it's
zero-indexed.

Frames have been converted to pointers so that we can update fields within
an instance.