Prevent change memory config without hard reset
I found out that `Reset()` only clears some of the registers. If you set
`Config.BlinkAndPWMSets` to `6`, then subsequently change it to `1`, writing
to frames 7 and beyond will result in empty frames. So warn that you need to
hard reset the device (power off or trigger reset pin).

To check this we need to read the existing config from the device before
setting it. I might implement read methods for each of the registers in
future, but for now we can extract the bits that we need.